extract index.js
This refactors the code to improve readability and reusability.
